A3 Finished

+++ server/src/routes/memories.ts
@@ -3,8 +3,20 @@ import { prisma } from '../lib/prisma';
 import { z } from 'zod'
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get('/memories', async () => {
+  app.get('/memories', async (request) => {
+    /* PARA QUANDO FOR NECESSARIO AUTENTICACAO NAS ROTAS, USAMOS O REQUEST NA LINHA 6 
+    ELA VERIFICA SE O TOKEN ESTA VINDO NA REQUISICAO PARA PODER LIBERAR O USO DA ROTA.
+    */
+    /*PREHANDLER - PARA USAR O HOOK EM CADA REQUISICAO E PASSANDO O JWTVERIFY COMO UM MIDDLEWARE.*/
+
+    app.addHook('preHandler', async (request) => {
+      await request.jwtVerify()
+    })
+
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub
+      },
       orderBy: {
         createdAt: 'asc'
       }
@@ -18,7 +30,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
     })
   })
 
-  app.get('/memories/:id', async (request) => {
+  app.get('/memories/:id', async (request, reply) => {
 
     const paramsSchema = z.object({
       id: z.string().uuid()
@@ -31,6 +43,11 @@ export async function memoriesRoutes(app: FastifyInstance) {
         id
       }
     })
+
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
+
     return memory
   })
 
@@ -48,13 +65,13 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl: '',
         isPublic,
-        userId: '79c7b9b5-174a-482f-9d17-38418c222791',
+        userId: request.user.sub,
       }
     })
     return memory
   })
 
-  app.put('/memories/:id', async (request) => {
+  app.put('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid()
     })
@@ -69,7 +86,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      }
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401)
+    }
+
+    memory = await prisma.memory.update({
       where: {
         id
       },
@@ -82,14 +109,24 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return memory
   })
 
-  app.delete('/memories/:id', async (request) => {
+  app.delete('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid()
     })
 
     const { id } = paramsSchema.parse(request.params)
 
-    const memory = await prisma.memory.delete({
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id
+      }
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401)
+    }
+
+    await prisma.memory.delete({
       where: {
         id
       }
